Add '-f' flag to be used with 'exec' in rebase interactive
Under rebase-interactive's exec statement, 'git-bottle' normally refused
to work because there is a background Git process still waiting, holding
state, and possiblly locking things.

And as we can see, with the error we are getting when we add a 'exec
git-bottle -f' line:

    fatal: update_ref failed for ref 'refs/heads/master': cannot lock ref 'refs/heads/master': ref refs/heads/master is at 6ca2a2e2299baf18247834ee670e1f6ee13efda5 but expected 801821e5855cbe7461f681ff2b36d1d688ed1761
    Successfully rebased and updated refs/heads/master.

Despite the error, 'git-bottle' returns of a zero exit sstatus,
'git-unbottle' works, and we found oursleves back in the state of the
rebase without the specific 'exec' line that included the call to
'git-bottle', so it is possible to do 'rebase --continue' from that
point.